Move debounce logic out of windows mod
